Refactor FlowHandle creation into FlowStateMachine.
