+++ core/src/main/kotlin/net/corda/core/flows/FlowStateMachine.kt
@@ -4,6 +4,7 @@ import co.paralleluniverse.fibers.Suspendable
 import com.google.common.util.concurrent.ListenableFuture
 import net.corda.core.crypto.Party
 import net.corda.core.crypto.SecureHash
+import net.corda.core.messaging.FlowProgressHandle
 import net.corda.core.node.ServiceHub
 import net.corda.core.serialization.CordaSerializable
 import net.corda.core.transactions.SignedTransaction
@@ -40,6 +41,8 @@ interface FlowStateMachine<R> {
 
     @Suspendable
     fun waitForLedgerCommit(hash: SecureHash, sessionFlow: FlowLogic<*>): SignedTransaction
+
+    fun createHandle(hasProgress: Boolean): FlowProgressHandle<R>
 
     val serviceHub: ServiceHub
     val logger: Logger
