+++ node/src/main/kotlin/net/corda/node/services/statemachine/FlowStateMachineImpl.kt
@@ -6,6 +6,7 @@ import co.paralleluniverse.fibers.Suspendable
 import co.paralleluniverse.strands.Strand
 import com.google.common.util.concurrent.ListenableFuture
 import com.google.common.util.concurrent.SettableFuture
+import net.corda.client.rpc.notUsed
 import net.corda.core.abbreviate
 import net.corda.core.crypto.Party
 import net.corda.core.crypto.SecureHash
@@ -13,7 +14,9 @@ import net.corda.core.flows.FlowException
 import net.corda.core.flows.FlowLogic
 import net.corda.core.flows.FlowStateMachine
 import net.corda.core.flows.StateMachineRunId
+import net.corda.core.messaging.FlowProgressHandle
 import net.corda.core.random63BitValue
+import net.corda.core.serialization.CordaSerializable
 import net.corda.core.transactions.SignedTransaction
 import net.corda.core.utilities.ProgressTracker
 import net.corda.core.utilities.UntrustworthyData
@@ -27,6 +30,7 @@ import org.jetbrains.exposed.sql.Transaction
 import org.jetbrains.exposed.sql.transactions.TransactionManager
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
+import rx.Observable
 import java.sql.Connection
 import java.sql.SQLException
 import java.util.*
@@ -97,6 +101,12 @@ class FlowStateMachineImpl<R>(override val id: StateMachineRunId,
     init {
         logic.stateMachine = this
     }
+
+    override fun createHandle(hasProgress: Boolean): FlowProgressHandle<R> = FlowProgressHandleImpl(
+            id = id,
+            returnValue = resultFuture,
+            progress = if (hasProgress) logic.track()?.second ?: Observable.empty() else Observable.empty()
+    )
 
     @Suspendable
     override fun run() {
@@ -408,5 +418,21 @@ class FlowStateMachineImpl<R>(override val id: StateMachineRunId,
         // Start time gets serialized along with the fiber when it suspends
         val duration = System.nanoTime() - startTime
         timer.update(duration, TimeUnit.NANOSECONDS)
+    }
+}
+
+@CordaSerializable
+private data class FlowProgressHandleImpl<A> (
+        override val id: StateMachineRunId,
+        override val returnValue: ListenableFuture<A>,
+        override val progress: Observable<String>) : FlowProgressHandle<A> {
+
+    /**
+     * Use this function for flows that returnValue and progress are not going to be used or tracked, so as to free up server resources.
+     * Note that it won't really close if one subscribes on progress [Observable], but then forgets to unsubscribe.
+     */
+    override fun close() {
+        progress.notUsed()
+        returnValue.cancel(false)
     }
 }
